extras: Enable direct-io options in group virt
Direct-io options are currently enabled in the ovirt-gluster
installation through cockpit ui whereas all the other optimizations in
gluster for virt are enabled through the group-virt option using
'volume-set' command. For the sake of completeness and to avoid any
confusion, it would be better to set the o-direct gluster options also
as part of group virt.

Change-Id: Iac4c443c2a8239effeb05a52344e563f1d877916
Fixes: bz#1758984
Signed-off-by: Krutika Dhananjay <kdhananj@redhat.com>