performance/read-ahead: update expected offset before unwinding read response
With the current code there is a window of time between unwinding
response to a read request and internal offset is updated to account
the read just done. If new sequential read request comes in this time
window, it is incorrectly identified as non-sequential read.

Fix is to update the file offset to account for a read request before
sending back the response to it.

Change-Id: Iff0c59c769e1eb15f262257763026657e2d4785d
Signed-off-by: Raghavendra G <rgowdapp@redhat.com>
Fixes: bz#1753843