posix-metadata.c: try to getxattr in one call.
Another location where instead of 2 sys calls we strive to get the
xattr in a single call, by guesstimating the required size
And avoid (or try to) not to first read the xattr len,
then another call to actually fetch. Instead, use a sane size
(256 bytes - worth checking if it makes sense or by default
use a larger size), and see if we can fetch it.
If we fail, we'll read the size and re-fetch.

Such changes are needed elsewhere too (see
https://github.com/gluster/glusterfs/issues/720 )

Change-Id: I466cea9d8b12fc45f6b37d202b1294ca28cd1fdd
updates: bz#1193929
Signed-off-by: Yaniv Kaul <ykaul@redhat.com>