errorcode: more information as part of return code
The details of how to achieve this is described in the issue, please
refer. Idea is, utilze those unused 31bits in 'op_ret' (ie, return
value) field for identifying the exact error (or as close as possible).

Plan is to handle fop by fop for this change. Best way to see the result
is by changing the lookup() fop, as that actually is expected to return
failure most times than success.

Updates: #280

Change-Id: If6e22e7aa170f3800c0f29156c4bfc700f790fef
Signed-off-by: Amar Tumballi <amar@kadalu.io>